refactor create xml from array

+++ src/helpers.php
@@ -74,7 +74,7 @@ if (!function_exists('get_random_string')) {
      */
     function get_random_string(int $length = 16): string
     {
-        return substr(str_shuffle(str_repeat($x='0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', ceil($length/strlen($x)) )), 1, $length);
+        return substr(str_shuffle(str_repeat($x='0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', (int) ceil($length/strlen($x)) )), 1, $length);
     }
 }
